RoutineList: Days of the week, task-symbols, routines total duration

+++ app/src/main/java/com/example/iteratine/Task.kt
@@ -3,7 +3,7 @@ package com.example.iteratine
 data class Task (
     val id: Int,
     val name: String,
-    val duration: Long,
+    val duration: Int,
     val emoji: String,
     val note: String,
     val parentRoutineID: Int,
